fix/payments data add in database
- add payments data from funstion where subscription or payment created
- and update status by webhooks events of payment intent succeeded and
  failed
- add custom error handling for check user exist , payment method attached to
  customer when they try to purchase
- [WIP] throw custom stripe errors from backend
- change all dates to iso string for add to db
- add every purchase details when subscription interval invoices are
  paid after interval
- create webhook helper function for making code easily readable and
  short

+++ components/paymentMethod/paymentMethod.service.js
@@ -1,5 +1,6 @@
 const userDb = require("../../dbUtils/userDb");
 const stripeHelper = require("../../utils/stripeHelper");
+const _ = require("lodash");
 class PaymentMethodService {
   async createPaymentMethod(type, paymentDetails, billingDetails, customerId) {
     try {
@@ -30,9 +31,12 @@ class PaymentMethodService {
         paymentMethod.type,
         customerId
       );
-      return { status: true };
+      return { paymentMethod };
     } catch (err) {
-      console.log({ "something goes wrong while create payment method": err });
+      console.log({ "Error from create payment method": err });
+      // if (err.type.includes("Stripe")) {
+      //   stripeHelper.throwStripeErrors(err);
+      // }
       throw new Error(err.message);
     }
   }
@@ -45,13 +49,16 @@ class PaymentMethodService {
       );
       return { status: true };
     } catch (err) {
-      console.log({ "something goes wrong while update payment method": err });
+      console.log({ "Error from update payment method": err });
+      // if (err.type.includes("Stripe")) {
+      //   stripeHelper.throwStripeErrors(err);
+      // }
       throw new Error(err.message);
     }
   }
   async getAllPaymentMethodsOfCustomer(stripeCustomerId) {
     try {
-      const paymentMethods = await stripeHelper.getAllPaymentMethodsById(
+      const paymentMethods = await stripeHelper.getAllPaymentMethodsByCutomerId(
         stripeCustomerId
       );
 
@@ -74,9 +81,10 @@ class PaymentMethodService {
       }
       return { paymentMethods: paymentMethods.data };
     } catch (err) {
-      console.log({
-        "something goes wrong while getting all payment methods ": err,
-      });
+      console.log({ "Error from get all payment method": err });
+      // if (err.type.includes("Stripe")) {
+      //   stripeHelper.throwStripeErrors(err);
+      // }
       throw new Error(err.message);
     }
   }
@@ -101,14 +109,31 @@ class PaymentMethodService {
         status: true,
       };
     } catch (err) {
-      console.log({
-        "something goes wrong while setting default payment method": err,
-      });
+      console.log({ "Error from setting default payment method": err });
+      // if (err.type.includes("Stripe")) {
+      //   stripeHelper.throwStripeErrors(err);
+      // }
       throw new Error(err.message);
     }
   }
   async deletePaymentMethodOfCustomer(stripeCustomerId, paymentMethodId) {
     try {
+      const subscriptionsList = await stripeHelper.getSubscriptionByCustomerId(
+        stripeCustomerId
+      );
+      const usedPaymentMethod = _.map(
+        subscriptionsList,
+        "default_payment_method" // get only payment methods id from array by lodash map method
+      );
+      if (usedPaymentMethod.includes(paymentMethodId)) {
+        throw new Error("CANNOT_DELETE_METHOD");
+      }
+      let customerPaymentMethods =
+        await stripeHelper.getAllPaymentMethodsByCutomerId(stripeCustomerId);
+      customerPaymentMethods = _.map(customerPaymentMethods.data, "id"); // get only payment methods id from array by lodash map method
+      if (!customerPaymentMethods.includes(paymentMethodId)) {
+        throw new Error("PAYMENT_METHOD_NOT_ATTACHED");
+      }
       await stripeHelper.detachMethodFromCustomer(paymentMethodId);
       await userDb.deletePaymentMethodFromCustomerDB(
         stripeCustomerId,
@@ -118,7 +143,10 @@ class PaymentMethodService {
         status: true,
       };
     } catch (err) {
-      console.log({ "something goes wrong while delete payment method": err });
+      console.log({ "Error from delete payment method": err });
+      // if (err.type.includes("Stripe")) {
+      //   await stripeHelper.throwStripeErrors(err);
+      // }
       throw new Error(err.message);
     }
   }
